13:43 8/4/21

+++ playerField.h
@@ -6,7 +6,7 @@ extern int player1Field[BOARD_HEIGHT][BOARD_WIDTH];
 extern int Level;
 extern int Line;
 extern int Point;
-extern int Time;
+extern double Time;
 
 void printField();
 void unite(Block& block);
